+++ packages/component-library/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## v9.3.0 (July 22, 2026)
+
+### Added
+
+- Added Horizontal Stacked Bar component.
+
+### Fixed
+
+- Resolved security vulnerabilities reported by GitHub.
+
 ## v9.2.1 (March 25, 2026)
 
 ### Changed